+++ frontend/src/hooks/index.ts
@@ -20,6 +20,7 @@ export { useRoadmapForm } from './useRoadmapForm';
 export { useDashboard } from './useDashboard';
 export { useBadgeNotifier } from './useBadgeNotifier';
 export { useLearningLogs, useLearningLogCalendar, useStreak } from './useLearningLogs';
+export { useLearningLogForm } from './useLearningLogForm';
 export { useSnippetComments } from './useSnippetComments';
 export { useAdvice, useAIChat, useConversations } from './useAdvice';
 export { usePomodoroTimer } from './usePomodoroTimer';
